Switch to class for embedding metadata
